ci: use a specific hcloud binary version
Using the "latest" version might run into the issue that a new CLI
version is being released, the tag is created, but the binary is not
yet attached.

In this case, the action just fails and the server keeps running, and
the server would still be billed.

Using a specific version should work around this issue.